build: rename to secure-store
Resolves: DCMAW-7544
