+++ buildLogic/plugins/src/main/kotlin/uk/gov/securestore/ext/ProjectExt.kt
@@ -1,4 +1,4 @@
-package uk.gov.securestorage.ext
+package uk.gov.securestore.ext
 
 import org.gradle.api.Project
