+++ buildLogic/plugins/src/main/kotlin/uk/gov/securestore/filetree/fetcher/FileTreesFetcher.kt
@@ -1,9 +1,9 @@
-package uk.gov.securestorage.filetree.fetcher
+package uk.gov.securestore.filetree.fetcher
 
 import org.gradle.api.Project
 import org.gradle.api.file.FileTree
 import org.gradle.api.provider.Provider
-import uk.gov.securestorage.ext.ProjectExt.debugLog
+import uk.gov.securestore.ext.ProjectExt.debugLog
 
 /**
  * Decorator class for containing multiple [FileTreeFetcher] objects.
