+++ buildLogic/plugins/src/main/kotlin/uk/gov/securestore/jacoco/config/JacocoUnitTestConfig.kt
@@ -1,14 +1,13 @@
-package uk.gov.securestorage.jacoco.config
+package uk.gov.securestore.jacoco.config
 
-import com.android.build.gradle.internal.tasks.DeviceProviderInstrumentTestTask
+import com.android.build.gradle.tasks.factory.AndroidUnitTest
 import org.gradle.api.Project
 import org.gradle.api.file.FileTree
-import org.gradle.api.tasks.TaskProvider
 import org.gradle.kotlin.dsl.named
-import uk.gov.securestorage.filetree.fetcher.FileTreeFetcher
+import uk.gov.securestore.filetree.fetcher.FileTreeFetcher
 
 /**
- * A [JacocoCustomConfig] implementation specifically for instrumentation tests.
+ * A [JacocoCustomConfig] implementation specifically for unit tests.
  *
  * Obtains the required Gradle task via the provided [capitalisedVariantName], passed back through
  * the [getExecutionData] function.
@@ -20,7 +19,7 @@ import uk.gov.securestorage.filetree.fetcher.FileTreeFetcher
  * @param capitalisedVariantName The TitleCase representation of the Android build variant of the
  * Gradle module. Finds the relevant test task name by using this parameter.
  */
-class JacocoConnectedTestConfig(
+class JacocoUnitTestConfig(
     private val project: Project,
     private val classDirectoryFetcher: FileTreeFetcher,
     private val capitalisedVariantName: String,
@@ -30,22 +29,19 @@ class JacocoConnectedTestConfig(
 ) {
 
     override fun getExecutionData(): FileTree {
-        val connectedTestTask: TaskProvider<DeviceProviderInstrumentTestTask> = project.tasks.named(
-            "connected${capitalisedVariantName}AndroidTest",
-            DeviceProviderInstrumentTestTask::class,
+        val unitTestTask = project.tasks.named(
+            "test${capitalisedVariantName}UnitTest",
+            AndroidUnitTest::class,
         )
-        val connectedTestExecutionDirectory = connectedTestTask.flatMap { connectedTask ->
+        val unitTestExecutionDataFile = unitTestTask.flatMap { utTask ->
             project.provider {
-                connectedTask
-                    .coverageDirectory
-                    .asFile
-                    .get()
+                utTask.jacocoCoverageOutputFile.get().asFile
+                    .parentFile
                     .absolutePath
             }
         }
-
-        return project.fileTree(connectedTestExecutionDirectory) {
-            setIncludes(listOf("**/*.ec"))
+        return project.fileTree(unitTestExecutionDataFile) {
+            setIncludes(listOf("${unitTestTask.name}.exec"))
         }
     }
 }
